+++ src/lnav_commands.cc
@@ -2347,7 +2347,7 @@ static string com_toggle_field(exec_context &ec, string cmdline, vector<string> 
             bool hide = args[0] == "hide-fields";
             vector<string> found_fields, missing_fields;
 
-            for (int lpc = 1; lpc < args.size(); lpc++) {
+            for (int lpc = 1; lpc < (int)args.size(); lpc++) {
                 intern_string_t name;
                 log_format *format = nullptr;
                 size_t dot;
